Use TDSVER in env to set the TDS version
Currently, version switching via the dbsetlversion() API uses legacy
Sybase constants, that have nothing to do with the TDS version strings,
and also is missing the constant to set the version to 5.0, required by
Sybase.

On the other hand, in the FreeTDS there's a private tds_set_verstr()
API that parses the contents of the TDSVER environment variable to set
the correct version ID in the login structure.

This way, plain "4.2", "7.1", "5.0" strings can be passed to the TinyTDS
Client constructor, and all versions are supported.